Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -6,8 +6,8 @@ char* welcome() {
 
 int main() {
     // printf("%s", welcome());
-    float c;
+    double c;
     scanf("%f",&c);
-    printf("you entered: %.2f",c);
+    printf("you entered: %f",c);
     return 0;
 }
